cleanup TestRunnerManagerTests
